devfreq: Allow zero values in opp table
commit ab8f58ad72c4d1abe59216362ddb8bfa428c9071 ("PM / devfreq: Set
min/max_freq when adding the devfreq device") introduces a change
where an error in finding the ceil or floor is returned as a
zero frequency. This return of zero is treated as error condition
from the callsites.

It is perfectly valid for a device to have a zero frequency, usually
it denotes a power collapse. This valid return of zero frequency ends
up being treated as an error condition.

Fix it.

Signed-off-by: Divyanshu-Modi <divyan.m05@gmail.com>
Change-Id: I44b4dd70179c89f210793b986bff430b22365635
(cherry picked from commit c404741f0947d8a1d0655965f3a58e80b366ed34)
Signed-off-by: TogoFire <togofire@mailfence.com>