do_vfs_ioctl(): use saner types
casting to pointer to int, only to pass that to function that
takes pointer to void and uses it as pointer to structure is
really asking for trouble.

"Some pointer, I'm not sure what to" is spelled "void *",
not "int *"; use that.

And declare the functions we are passing that pointer to
as taking the pointer to what they really want to access.

Signed-off-by: Al Viro <viro@zeniv.linux.org.uk>
Change-Id: Ia8bde3b4b85c0703489ac9b89490d49b18c0844b
Signed-off-by: Arnd Bergmann <arnd@arndb.de>
Signed-off-by: Divyanshu-Modi <divyan.m05@gmail.com>
Signed-off-by: Divyanshu-Modi <divyan.m05@gmail.com>
(cherry picked from commit 4d6ac27452c16bc3d2ba3a24c440060cf519653c)
Signed-off-by: TogoFire <togofire@mailfence.com>