crypto: x86/chacha - expose SIMD ChaCha routine as library function
commit 84e03fa39fbe95a5567d43bff458c6d3b3a23ad1 upstream.

Wire the existing x86 SIMD ChaCha code into the new ChaCha library
interface, so that users of the library interface will get the
accelerated version when available.

Given that calls into the library API will always go through the
routines in this module if it is enabled, switch to static keys
to select the optimal implementation available (which may be none
at all, in which case we defer to the generic implementation for
all invocations).

Signed-off-by: Ard Biesheuvel <ardb@kernel.org>
Change-Id: Ibb1d270e1c5204b959bdfe33c0f9c3c2187473e5
Signed-off-by: Herbert Xu <herbert@gondor.apana.org.au>
Signed-off-by: Jason A. Donenfeld <Jason@zx2c4.com>
(cherry picked from commit 910eb3cc6e7ab7b92dc3322c6900cc10922dbe52)
Signed-off-by: TogoFire <togofire@mailfence.com>