crypto: arm64/chacha - expose arm64 ChaCha routine as library function
commit b3aad5bad26a01a4bd8c49a5c5f52aec665f3b7c upstream.

Expose the accelerated NEON ChaCha routine directly as a symbol
export so that users of the ChaCha library API can use it directly.

Given that calls into the library API will always go through the
routines in this module if it is enabled, switch to static keys
to select the optimal implementation available (which may be none
at all, in which case we defer to the generic implementation for
all invocations).

Signed-off-by: Ard Biesheuvel <ardb@kernel.org>
Change-Id: I55e3222119a107f3bfd76d85a9ce79bec656fec3
Signed-off-by: Herbert Xu <herbert@gondor.apana.org.au>
Signed-off-by: Jason A. Donenfeld <Jason@zx2c4.com>
(cherry picked from commit 6302eefd8c9c12640989c49d2736edfd353d2c0b)
Signed-off-by: TogoFire <togofire@mailfence.com>