crypto: arm/chacha - expose ARM ChaCha routine as library function
commit a44a3430d71bad4ee56788a59fff099b291ea54c upstream.

Expose the accelerated NEON ChaCha routine directly as a symbol
export so that users of the ChaCha library API can use it directly.

Given that calls into the library API will always go through the
routines in this module if it is enabled, switch to static keys
to select the optimal implementation available (which may be none
at all, in which case we defer to the generic implementation for
all invocations).

Signed-off-by: Ard Biesheuvel <ardb@kernel.org>
Change-Id: I6baa1f5c3df5158ec99809f1e1d013c298665b9d
Signed-off-by: Herbert Xu <herbert@gondor.apana.org.au>
Signed-off-by: Jason A. Donenfeld <Jason@zx2c4.com>
(cherry picked from commit 237c477712f6351e578a831aab6baf993f09ce6c)
Signed-off-by: TogoFire <togofire@mailfence.com>