crypto: arm/curve25519 - import Bernstein and Schwabe's Curve25519 ARM implementation
commit f0fb006b604f98e2309a30f34ef455ac734f7c1c upstream.

This comes from Dan Bernstein and Peter Schwabe's public domain NEON
code, and is included here in raw form so that subsequent commits that
fix these up for the kernel can see how it has changed. This code does
have some entirely cosmetic formatting differences, adding indentation
and so forth, so that when we actually port it for use in the kernel in
the subsequent commit, it's obvious what's changed in the process.

This code originates from SUPERCOP 20180818, available at
<https://bench.cr.yp.to/supercop.html>.

Signed-off-by: Jason A. Donenfeld <Jason@zx2c4.com>
Change-Id: If19eb6ba117d0768698d5161a6afd28f7f1f356c
Signed-off-by: Ard Biesheuvel <ardb@kernel.org>
Signed-off-by: Herbert Xu <herbert@gondor.apana.org.au>
Signed-off-by: Jason A. Donenfeld <Jason@zx2c4.com>
Change-Id: If19eb6ba117d0768698d5161a6afd28f7f1f356c
(cherry picked from commit 59b7654e0f95eb4e1e988a343551a8bfeb054667)
Signed-off-by: TogoFire <togofire@mailfence.com>