wireguard: receive: remove dead code from default packet type case
commit 2b8765c52db24c0fbcc81bac9b5e8390f2c7d3c8 upstream.

The situation in which we wind up hitting the default case here
indicates a major bug in earlier parsing code. It is not a usual thing
that should ever happen, which means a "friendly" message for it doesn't
make sense. Rather, replace this with a WARN_ON, just like we do earlier
in the file for a similar situation, so that somebody sends us a bug
report and we can fix it.

Reported-by: Fabian Freyer <fabianfreyer@radicallyopensecurity.com>
Signed-off-by: Jason A. Donenfeld <Jason@zx2c4.com>
Change-Id: I7f4aeaa0560045e5f75a2c7796b6202fc67a970e
Signed-off-by: David S. Miller <davem@davemloft.net>
Signed-off-by: Jason A. Donenfeld <Jason@zx2c4.com>
Change-Id: I7f4aeaa0560045e5f75a2c7796b6202fc67a970e
(cherry picked from commit 8f3c8ba1fd1e5fd3b4abb93af863622661cd450f)
Signed-off-by: TogoFire <togofire@mailfence.com>