wireguard: send/receive: use explicit unlikely branch instead of implicit coalescing
commit 243f2148937adc72bcaaa590d482d599c936efde upstream.

It's very unlikely that send will become true. It's nearly always false
between 0 and 120 seconds of a session, and in most cases becomes true
only between 120 and 121 seconds before becoming false again. So,
unlikely(send) is clearly the right option here.

What happened before was that we had this complex boolean expression
with multiple likely and unlikely clauses nested. Since this is
evaluated left-to-right anyway, the whole thing got converted to
unlikely. So, we can clean this up to better represent what's going on.

The generated code is the same.

Suggested-by: Sultan Alsawaf <sultan@kerneltoast.com>
Signed-off-by: Jason A. Donenfeld <Jason@zx2c4.com>
Change-Id: Ib144e0222c150f82c56ec6d214395da6e1d70d38
Signed-off-by: David S. Miller <davem@davemloft.net>
Signed-off-by: Jason A. Donenfeld <Jason@zx2c4.com>
Change-Id: Ib144e0222c150f82c56ec6d214395da6e1d70d38
(cherry picked from commit 45894ecebf5f3552b3b0b854c1138081aceecf7c)
Signed-off-by: TogoFire <togofire@mailfence.com>