interconnect: Add helpers for enabling/disabling a path
There is a repeated pattern in multiple drivers where they want to switch
the bandwidth between zero and some other value. This is happening often
in the suspend/resume callbacks. Let's add helper functions to enable and
disable the path, so that callers don't have to take care of remembering
the bandwidth values and handle this in the framework instead.

With this patch the users can call icc_disable() and icc_enable() to lower
their bandwidth request to zero and then restore it back to it's previous
value.

Suggested-by: Evan Green <evgreen@chromium.org>
Suggested-by: Bjorn Andersson <bjorn.andersson@linaro.org>
Signed-off-by: Georgi Djakov <georgi.djakov@linaro.org>
Change-Id: Ia9245bbabbecb600a1004cd880e4e7a02b74b5c5
Reviewed-by: Matthias Kaehlcke <mka@chromium.org>
Link: https://lore.kernel.org/r/20200507120846.8354-1-georgi.djakov@linaro.org
Signed-off-by: Georgi Djakov <georgi.djakov@linaro.org>
Change-Id: Ia9245bbabbecb600a1004cd880e4e7a02b74b5c5
Signed-off-by: Divyanshu-Modi <divyan.m05@gmail.com>
(cherry picked from commit 5cbffddf50001cbaeaa21c7712966d1baa983180)
Signed-off-by: TogoFire <togofire@mailfence.com>