wireguard: selftests: use newer iproute2 for gcc-10
commit ee3c1aa3f34b7842c1557cfe5d8c3f7b8c692de8 upstream.

gcc-10 switched to defaulting to -fno-common, which broke iproute2-5.4.
This was fixed in iproute-5.6, so switch to that. Because we're after a
stable testing surface, we generally don't like to bump these
unnecessarily, but in this case, being able to actually build is a basic
necessity.

Signed-off-by: Jason A. Donenfeld <Jason@zx2c4.com>
Change-Id: Ia6edbfaff8e493bc03c4a350994dcee9740fd1a8
Signed-off-by: David S. Miller <davem@davemloft.net>
Signed-off-by: Jason A. Donenfeld <Jason@zx2c4.com>
Change-Id: Ia6edbfaff8e493bc03c4a350994dcee9740fd1a8
(cherry picked from commit 5842e9e2879fef40ec745257b60d57f66f3d9364)
Signed-off-by: TogoFire <togofire@mailfence.com>