lto: Allow LTO for sys_ni.c
At some earlier point LTO was disabled for sys_ni.c to handle the system
call aliases without problems. This is not needed anymore, now that we
express syscall aliaseing in C instead of assembler. In fact
current LTO builds require this file to be LTOed because of some strange
external symbol weak resolution quirks in gcc 10.

Signed-off-by: Andi Kleen <ak@linux.intel.com>
Change-Id: Ie1bd87baafb5136283ca346599f52a00a6e3e426
Signed-off-by: Vaisakh Murali <mvaisakh@statixos.com>
Signed-off-by: TogoFire <togofire@mailfence.com>