Kbuild, lto: Generate debuginfo for temporary kallsyms links
gcc 10 changes the layout of the data segment when --strip-debug is
specified with LTO. This causes problems with kallsyms, which builds
temporary vmlinux with --strip-debug, and expects the final vmlinux
without --strip-debug to have the same symbol table.

To work around this problem generate debuginfo the temporary
links too when LTO is enabled. This will cause some extra IO.

Signed-off-by: Andi Kleen <ak@linux.intel.com>
Change-Id: I80dff3c6341e3f6404f95c414d98e0fba7c3bd34
Signed-off-by: Vaisakh Murali <mvaisakh@statixos.com>
Signed-off-by: TogoFire <togofire@mailfence.com>