BACKPORT: lib: Add zstd support to decompress
- Add unzstd() and the zstd decompress interface.

- Add zstd support to decompress_method().

The decompress_method() and unzstd() functions are used to decompress
the initramfs and the initrd. The __decompress() function is used in
the preboot environment to decompress a zstd compressed kernel.

The zstd decompression function allows the input and output buffers to
overlap because that is used by x86 kernel decompression.

Tested-by: Sedat Dilek <sedat.dilek@gmail.com>
Reviewed-by: Kees Cook <keescook@chromium.org>
Link: https://lore.kernel.org/r/20200730190841.2071656-3-nickrterrell@gmail.com
[@ElectroPerf - Backport to 4.19]
Signed-off-by: Kunmun <kunmun.devroms@gmail.com>
Change-Id: Ifa0e40cb33de8f232992c9624e14a11e247c2980
Signed-off-by: Tashfin Shakeer Rhythm <tashfinshakeerrhythm@gmail.com>
(cherry picked from commit b0c1f91e189fbf25b02f97f2325b41906ca839ed)
Signed-off-by: TogoFire <togofire@mailfence.com>