BACKPORT: lib: decompress_unzstd: Limit output size
The zstd decompression code, as it is right now, will most likely fail
on 32-bit systems, as the default output buffer size causes the buffer's
end address to overflow.

Address this issue by setting a sane default to the default output size,
with a value that won't overflow the buffer's end address.

Reviewed-by: Nick Terrell <terrelln@fb.com>
[@ElectroPerf - Backport to 4.19]
Signed-off-by: Kunmun <kunmun.devroms@gmail.com>
Change-Id: I655f69d58322a5311774f9dfad53875aea91b485
Signed-off-by: Tashfin Shakeer Rhythm <tashfinshakeerrhythm@gmail.com>
Signed-off-by: TogoFire <togofire@mailfence.com>