build.config.msm.common: Treat MAKE_ARGS as an array
In kernel/build commit 7ff932f0ba96 ("build.sh: Treat MAKE_ARGS as an
array"), MAKE_ARGS becomes an array. Update build.config.msm.common
so that it also treats MAKE_ARGS as an array. This is
backwards-compatible with an older kernel/build and it's expanded the
same way if MAKE_ARGS is just a plain variable (not an array).

Signed-off-by: Elliot Berman <eberman@codeaurora.org>
Change-Id: I40f2616bd3afa78e3164e1960d32c2d26b1b3ff1
Signed-off-by: TogoFire <togofire@mailfence.com>