msm: kgsl: Fix syncsource spinlock recursion
Syncsource spinlocks can be taken in IRQ context, so it needs to
be IRQ-safe else it would lead to spinlock recursion. Disable the
interrupt when the lock is taken in interrupt context, and
re-enable it when while unlocking.

Signed-off-by: Nitheesh Muthuraj <nmuthuraj@codeaurora.org>
Change-Id: I4ea759562abe902998bdc979bb57bbd6473e80ee
(cherry picked from commit 4240d51f9e13c447f9317e52ad77a27150a253a0)
Signed-off-by: TogoFire <togofire@mailfence.com>