cpufreq: qcom-hw: Switch to non deferrable work
Currently, when lmh interrupt fires we notify the scheduler about
thermal pressure. The scheduler reduces the capacity of the cpu(s)
for which the interrupt fired, while we requeue our deferrable work.

This leads to an interesting deadlock, since the cpu is running with
reduced capacity, the scheduler does not put any task on it,
extending the idle time and causing it to remain at reduced capacity -
leading to severe underutilization of the cpu(s).

Switch to using delayed work instead of deferrable work. This will cause
us to exit idle until the thermal condition is recovered.

Signed-off-by: Abhijeet Dharmapurikar <adharmap@codeaurora.org>
Change-Id: I717b02270afa0dd66143f733c53be9436ea03827
Signed-off-by: Pranav Vashi <neobuddy89@gmail.com>
(cherry picked from commit 3a5dbee66c05991549e9a1230b3e049afc651249)
Signed-off-by: TogoFire <togofire@mailfence.com>