msm: kgsl: Fix sysfs deadlock in kgsl
If kgsl mem entries are being accessed in sysfs and if putting
back the mementry refcount triggers a free of the process private,
we have a mutex deadlock. Because freeing the process private triggers
removal of the sysfs directories from within a thread that is accessing
the sysfs files. Fix this by taking an extra refcount on the process
private and then putting it back in a deferred manner.

(cherry picked from commit 3458c4e5f3fd944adfd7bf23409358fe8a27ec5d)
Signed-off-by: TogoFire <togofire@mailfence.com>
Change-Id: Ibdafcdc6d72339d81c759608e5d9140d97461b31