techpack: disp: dsi: add check for any queued DSI CMDs before clock force update
During a force update of DSI clocks, the state of the byte and pclks are
toggled irrespective of the ref-count. This in addition with ASYNC
command wait can result in interrupt storm, if and when the clocks are
being toggled a previous command that was triggered using the ASYNC
wait flag fires an ISR. The interrupt status doesn't get cleared if
the ISR is being serviced with the clocks are off.

The change adds a check for pending queued commands before any force
update of DSI clocks.

Signed-off-by: Satya Rama Aditya Pinapala <psraditya30@codeaurora.org>
Change-Id: Ia36606a0de52d985ab5c91d706c8ba5698008c9e
Signed-off-by: Rajeev Nandan <quic_rajeevny@quicinc.com>
Signed-off-by: Divyanshu-Modi <divyan.m05@gmail.com>
(cherry picked from commit 3403eb2c36b8de1864eaac7ab6487b31531414c3)
Signed-off-by: TogoFire <togofire@mailfence.com>