sched/walt: Increase the number of history samples
The number history samples should be raised to a higher value so that an
active history of 100ms is considered in accounting task demand. Raise
the history size from 5 to 8 given the default windows size is 8ms. This
will capture 64ms of history for 8ms windows, and 128ms of history for
16ms windows, granting a compromise to prevent further code bloat.

Signed-off-by: Shaleen Agrawal <quic_shalagra@quicinc.com>
Change-Id: If2b4b5e236518414104457fe26a7f8c3a25ad27c
Signed-off-by: Alexander Winkowski <dereference23@outlook.com>
(cherry picked from commit abf868670c73878409371257909cd313ab879bf4)
Signed-off-by: TogoFire <togofire@mailfence.com>