msm: kgsl: Avoid gpu boot if it is already up
Consider a scenario when touch wakeup happened before first open.
In such case, touch wakeup thread will turn on the GPU. Currently,
when first open happens, KGSL blindly goes ahead to turn on the GPU.
This will result in inappropriate system state. To avoid this, ensure
that KGSL does not attempt to turn on GPU if it is already on.

Signed-off-by: Kamal Agrawal <quic_kamaagra@quicinc.com>
Change-Id: I43c334d968ddb898bffb68dc4f25701e6ab6c2e4
(cherry picked from commit f4bcd5a9428639280107cbcf5150355528c34bf1)
Signed-off-by: TogoFire <togofire@mailfence.com>