ANDROID: remove stale variables from build.config files
There is no need for CROSS_COMPILE_COMPAT as of v5.16-rc1 via
commit 3e6f8d1fa184 ("arm64: vdso32: require CROSS_COMPILE_COMPAT for gcc+bfd")
when LLVM=1 is set. This was backported to v5.10.89.

LINUX_GCC_CROSS_COMPILE_PREBUILTS_BIN was a temporary dependency until
LLVM_IAS=1 was enabled for all architectures. With LLVM_IAS=1 (implied
by LLVM=1), nothing in this directory is used and the variable can be
removed.

It doesn't hurt to respecify any of the above, but they are no longer
necessary.

We may be able to further simplify the usage of LLVM_IAS=1 and
CROSS_COMPILE in android13-5.10 depending on whether the following get
backported in the future:

commit f12b034afeb3 ("scripts/Makefile.clang: default to LLVM_IAS=1").
commit 231ad7f409f1 ("Makefile: infer --target from ARCH for CC=clang")

Bug: 209655537
Signed-off-by: Nick Desaulniers <ndesaulniers@google.com>
Change-Id: I70e2111b254e567c0f3493a2eef17a993a98668e
(cherry picked from commit ba302abfca527e90ffc3cb75fe8c6f5504f25f85)
Signed-off-by: TogoFire <togofire@mailfence.com>