scsi: ufs-qcom: Remove current pm_qos implementation
These pieces of codes affine a pm_qos before a transfer request is
called, then disables it after the UFS clock is turned off, leaving the
CPU in a shallow idle state for longer than necessary. Furthermore, this
affines the pm_qos to all cores, wasting immense amounts of power.
Remove the code entirely in order to replace it with a different
implementation.

Change-Id: I213a4cb045f43fb4f65acde78bf2bee8b1130ed9
Signed-off-by: Kazuki Hashimoto <kazukih@tuta.io>
Signed-off-by: Adithya R <gh0strider.2k18.reborn@gmail.com>