sched: walt: Introduce bucket bitmask
In get_pred_busy, we attempt to find the closing bucket, which is
nothing but the first bucket which is filled after the start bucket.

We can optimize this computation by using a bitmask of buckets.

Signed-off-by: Shaleen Agrawal <quic_shalagra@quicinc.com>
Change-Id: I0e69156aea148b0441ceab0a26c0553c5f87f3e8
[dereference23: Backport to msm-5.4]
Signed-off-by: Alexander Winkowski <dereference23@outlook.com>
Signed-off-by: Tashfin Shakeer Rhythm <tashfinshakeerrhythm@gmail.com>
(cherry picked from commit 6202765379ade7dd4b765566df3a57d1d0d8d95b)
Signed-off-by: TogoFire <togofire@mailfence.com>