sched: walt: Remove finding historical runtimes in pred
When determining predictive demand, we currently look at the history
samples in our past 8 windows that fall under the bucket range for which
we are evaluating. However, this level of precision is not required, and
can be dropped in favor of reduction of overheads.

Signed-off-by: Shaleen Agrawal <quic_shalagra@quicinc.com>
Change-Id: I9ef363da3e6d0f20ec3999e2f4246c6a540bd875
Signed-off-by: Alexander Winkowski <dereference23@outlook.com>
Signed-off-by: Tashfin Shakeer Rhythm <tashfinshakeerrhythm@gmail.com>
(cherry picked from commit 96a99d1cbfdea9ea7a4c0b20a5acf9184b669feb)
Signed-off-by: TogoFire <togofire@mailfence.com>