bluebooth: Compile out debug logs when !CONFIG_DEBUG_KERNEL
Each time when phone is connected to bluetooth, a large amount
of unnecessary logs get dumped at the dmesg. Hence, guard the
define declaraion of BTFMSLIM_INFO with an if/else macro check of
CONFIG_DEBUG_KERNEL as the logs are only useful when debugging and
we aren't debugging the kernel right now.

Signed-off-by: Tashfin Shakeer Rhythm <tashfinshakeerrhythm@gmail.com>
Change-Id: Ibcfc611592b4d730a48ed1814289478984897296
Signed-off-by: TogoFire <togofire@mailfence.com>