lpm-levels: Ensure that timer pending is read before writing timeout val
It is possible for the block of code in the spinlock to be reordered,
causing RIMPS timer pending to be read after the writes are finished,
despite it having to be read prior to the writes. Fix it.

Signed-off-by: Kazuki Hashimoto <kazukih@tuta.io>
Change-Id: I820fd98fd0af0defb681f8f5c19cc6d341cc001b
Signed-off-by: Pranav Vashi <neobuddy89@gmail.com>
(cherry picked from commit 4be1b59adfe1db5f16bba3d5dad7e902126b25c0)
Signed-off-by: TogoFire <togofire@mailfence.com>