ANDROID: firmware_loader: Restore permissions for firmware class path
Permissions for accessing firmware class path was wrongly set to write only for root
as a result the path can not be read. Restore permissions, so that it can be read.
Add a getter function to read the class path.

Bug: 240542962
Fixes: d551647f3bce ("ANDROID: firmware_loader: Add support for customer firmware paths")
Fixes: 87abd99c7f22 ("ANDROID: firmware_loader: Fix warning with firmware_param_path_set").
Signed-off-by: Pranav Vashi <neobuddy89@gmail.com>
Change-Id: I70333b771f79426821e187aaa69f89e23dc435c6
(cherry picked from commit 9775c3008c2a319eb3c63147804afc0039a927cf)
Signed-off-by: TogoFire <togofire@mailfence.com>