LRNG - sysctls and /proc interface
The LRNG sysctl interface provides the same controls as the existing
/dev/random implementation. These sysctls behave identically and are
implemented identically. The goal is to allow a possible merge of the
existing /dev/random implementation with this implementation which
implies that this patch tries have a very close similarity. Yet, all
sysctls are documented at [1].

The sysctl implementation is only enabled if the existing random.c file
is not compiled.

Signed-off-by: Stephan Mueller <smueller@chronox.de>
Change-Id: I1e48a954f1a4664e7f99875acbff6561fcde8611
(cherry picked from commit 2f97e216e6f96cf46c20591464c8dcdaf979d7a9)
Signed-off-by: TogoFire <togofire@mailfence.com>