aw8697: Drop OPLUS_HAPTIC_OOS changes for setting vmax to 100-109
There is a weird discontinuity in vibration strength when changing vmax
from 90 to 100 and from 100 to 110. This looks to be some change related
to OOS haptics, but since we're not using that, we can just drop this
change.

(cherry picked from commit 7b89caecf4b23b4315523cdf2eb0cfe7329bf3a1)
Signed-off-by: TogoFire <togofire@mailfence.com>
Change-Id: I019d2223c4135ff5e41dfb2709e7f23847500dce