defconfig: Set CONFIG_IP6_NF_TARGET_HL
The CONFIG_IP6_NF_TARGET_HL flag is a Linux kernel configuration option that controls firewall target support for IPv6 header length (HL). The HL is a field in the IPv6 packet header that specifies the length of the header in 64-bit words.

When this flag is enabled in the kernel, it supports using the firewall target to make rules based on the value of the HL field of the IPv6 header. This allows network administrators to implement more precise and restrictive firewall policies based on the IPv6 packet header.

Signed-off-by: AnierinB <anierin@evolution-x.org>
Change-Id: I64bca04acde707f96e75cc29a9e95dfaf1dfa3b0
Signed-off-by: mukesh22584 <mks22584@gmail.com>
Signed-off-by: TogoFire <togofire@mailfence.com>