arm64: dts: Don't blank screen as part of LP1 cmd
We send the LP1 command as part of entering Always-On Display mode, but
we don't want that command to blank the screen. After all, the point of
AoD mode is to keep the screen on to display the time, notifications,
etc.

Signed-off-by: chandu078 <chandudyavanapelli03@gmail.com>
Change-Id: I0a392fba4dcf313791b514b53454be284757accb
Signed-off-by: Pranav Vashi <neobuddy89@gmail.com>
(cherry picked from commit b973829b1420da8d40418d2775eefce978c9bd97)
Signed-off-by: TogoFire <togofire@mailfence.com>