kbuild: deb-pkg: set version for linux-headers paths
As a result of the switch to dh_listpackages, $version is no longer set
when install_kernel_headers() is called.  This causes files in the
linux-headers deb package to be installed to a path with an empty
$version (e.g. /usr/src/linux-headers-/scripts/sign-file rather than
/usr/src/linux-headers-6.3.0-rc3/scripts/sign-file).

To avoid this, while continuing to use the version information from
dh_listpackages, pass $version from $package as the second argument
of install_kernel_headers().

Fixes: 36862e14e316 ("kbuild: deb-pkg: use dh_listpackages to know enabled packages")
Signed-off-by: Kevin Locke <kevin@kevinlocke.name>
Change-Id: Idba7579e2d52380b1f518eff81ba345ef0b1d1f5
Signed-off-by: Masahiro Yamada <masahiroy@kernel.org>
(cherry picked from commit 3ced71d273f8edf07bf01a831a49ca6b988e06b3)
Signed-off-by: TogoFire <togofire@mailfence.com>