Kbuild,lto: account for cross compiler when checking for LTO
The lto script was written with respect to x86 LTO, hence it relied
on direct vars for checking the binutils and other compiler tool
versions straight from host. Cross compilation was never accounted
for, which lead to checking of older binaries of the host, and hence
disabling LTO for the target architecture.

This patch accounts for the same and LTO cross compilation works as
intended.

Signed-off-by: Vaisakh Murali <mvaisakh@statixos.com>
Change-Id: Icd8ed03b3823ca5a7fd67e5e78a58f1c36518582
Signed-off-by: TogoFire <togofire@mailfence.com>