techpack/dataipa: affine ipa irqs to little CPUs
Some of the ipa irqs are run on CPU0 only and a few are run on CPU4.
Instead of utilizing the prime or gold clusters, affine these
completely to little cluster.
Little cluster can easily handle these interrupts and manage them
in a power efficient manner rather than handing these to the gold
clusters.

Signed-off-by: Vaisakh Murali <mvaisakh@statixos.com>
Change-Id: I6d16cb249fb1753ddea1701d15eb137dc21a8467
Signed-off-by: TogoFire <togofire@mailfence.com>