kernel: add a config for building with PGO
GCC can make use of PGO (Profile-Guided Optimization) to enhance the
efficiency and performance of the kernel. This technique involves
analyzing the code to gather data about frequently executed parts and
frequently called functions. The compiler then utilizes this data to
optimize the most vital parts of the code, which leads to significant
performance gains, especially for larger and more complex areas of
the kernel.

Signed-off-by: Vaisakh Murali <mvaisakh@statixos.com>
Change-Id: I599bde5f2dd17148913420c1f1bff0bed057efc3
Signed-off-by: TogoFire <togofire@mailfence.com>