Revert "rtc: control default alarm wake up capability"
Our RTC peripheral in PM8941 PMIC does not support disabling RTC alarm wakeup.
This property is enabled only on msm-5.4 and msm-5.15 at the moment for iot
and vm platforms which indicates that it is clearly not for our SDM660 mobile
platform. We are never going to enable "disable-alarm-wakeup" at dts. So, drop
this useless patch.

The same applies to the PM8350.

This reverts commit 268f4b8f3f276d99421d804a7b92a636b9d2ff51.

Signed-off-by: Tashfin Shakeer Rhythm <tashfinshakeerrhythm@gmail.com>
Change-Id: I9e2fbf27b704420514719b09f938ca6810426773
Signed-off-by: TogoFire <togofire@mailfence.com>