defconfig: Disable DEBUG_KERNEL
The CONFIG_DEBUG_KERNEL setting is a configuration option present in the Linux kernel that enables a series of debugging and error checking features. These extra features are useful for developers and systems engineers who need to analyze specific problems and debug the kernel.

However, the recommendation to disable the CONFIG_DEBUG_KERNEL option is for production environments, where the focus is on system stability and performance. There are a few reasons for this:

1. Performance: Enabling kernel debug mode consumes additional resources such as processing and memory. This can negatively affect overall system performance, especially on production servers where efficiency is crucial.

2. Security: Debug mode may introduce some additional vulnerabilities as extra checks may create security holes. Therefore, in environments where security is paramount, it is recommended to disable this setting to reduce potential risks.

3. Complexity: The CONFIG_DEBUG_KERNEL option makes the kernel more complex, with features that are designed to help with development and debugging. This can make the maintenance process more challenging and complicated for system administrators.

Overall, the recommendation to disable CONFIG_DEBUG_KERNEL in production environments is to ensure better performance, security, and system management. However, in development environments or specific debugging cases, it is worth enabling it to help diagnose problems and troubleshoot bugs.

Signed-off-by: TogoFire <togofire@mailfence.com>
Change-Id: Ic1e4c2f9665ab8f897d6bf8ce6862499064ff020