init: Kconfig: Disable audit by default
CONFIG_AUDIT in the Linux kernel is responsible for collecting and logging security events in the kernel. These events can be used for auditing purposes, such as detecting and investigating security breaches.

On production systems, the CONFIG_AUDIT option is generally disabled to improve performance. However, it is important to keep the CONFIG_AUDIT option enabled on development and test systems.

And drop in defconfig too.

Signed-off-by: TogoFire <togofire@mailfence.com>
Change-Id: I9fd5693263a07b9380e2a3a3c08a55242f3ccf87