kgsl: Update kgsl_cancel_event() to cancel the first matching event
When cancelling events, we go through the entire list and remove all
matching events. This is not what we want since callers expect to cancel
a single event. So, exit the search when we find and cancel the first
matching event in the list.

Signed-off-by: Lynus Vaz <quic_lvaz@quicinc.com>
Change-Id: Idb4bc8ab528eb054ebcdfa1ae645f3031bc6ce73
(cherry picked from commit 571279ae8bd4a39c788476b49f98a3239bef5bf2)
Signed-off-by: TogoFire <togofire@mailfence.com>