kgsl: Update kgsl_cancel_event() to cancel the first matching event
When cancelling events, we go through the entire list and remove all
matching events. This is not what we want since callers expect to cancel
a single event. So, exit the search when we find and cancel the first
matching event in the list.

Signed-off-by: Lynus Vaz <quic_lvaz@quicinc.com>
Change-Id: Iaa5c21e5719b80173275a7adb80a10800e5f6e9b
(cherry picked from commit 571279ae8bd4a39c788476b49f98a3239bef5bf2)
Signed-off-by: TogoFire <togofire@mailfence.com>