lpm-levels: Treat negative sleep lengths as indefinite
A negative sleep length means there's no hrtimer armed, and therefore the
sleep length is indefinite if the tick is stopped. Fix this logic in TEO to
allow a deeper idle state to correctly be used when there's no hrtimer
armed and thus the CPU won't be prematurely woken up from its sleep.

Change-Id: Id34f8ba930e70ba4cb6969bb780e2889d04da852
Signed-off-by: Sultan Alsawaf <sultan@kerneltoast.com>
[dereference23: Port to lpm-levels]
Signed-off-by: Alexander Winkowski <dereference23@outlook.com>
Signed-off-by: TogoFire <togofire@mailfence.com>