input: misc: add pwrkey-sync software driver
This software driver listens to power button events and performs sync when
it's held down for long.

This is to minimize data corruption when the system is not responding and
the user wants to initiate a forced reboot.

Additionally, it can forcefully trigger a kernel panic to help debugging a
kernel issue further by storing its log to something like pstore before the
hardware resets itself.

Signed-off-by: Juhyung Park <qkrwngud825@gmail.com>
Change-Id: Ieae3363eaa13574cc23e2c3d3871a0fbdcdc2dc6
Signed-off-by: TogoFire <togofire@mailfence.com>