input: misc: add pwrkey-sync software driver
This software driver listens to power button events and performs sync when
it's held down for long.

This is to minimize data corruption when the system is not responding and
the user wants to initiate a forced reboot.

Additionally, it can forcefully trigger a kernel panic to help debugging a
kernel issue further by storing its log to something like pstore before the
hardware resets itself.

Signed-off-by: Juhyung Park <qkrwngud825@gmail.com>
Change-Id: Ie4e2417f6581d615692873b2cb1f05721ef3a191
Signed-off-by: TogoFire <togofire@mailfence.com>