drivers: usb: dwc3: Nuke sync probe
Revert "drivers: usb: dwc3: Specify sync probe for dwc3 driver"
This reverts commit 2b30382e1e8aec5a65eea0fa21ae36180d2c21dd.

Revert "drivers: usb: dwc3: Specify sync probe for dwc3-of-simple."
This reverts commit a3514d484a91f6c73a96a5e35cb777cfb94cd517.

Revert "drivers: usb: dwc3: Specify sync probe for msm-dwc3 driver"
This reverts commit 2906886e3f23f80b014f268a664cb1f208473d3a.

Reason: Forced synchronization can negatively affect data transfer performance.

The "sync probe" is a technique used to ensure that the kernel driver is synchronized with the associated hardware. It forces the driver to check the current state of the hardware before starting the operation.

The primary purpose of the "sync probe" is not to reduce boot time, although it may have that effect as a by-product.
The "sync probe" primarily aims to ensure system stability, reliability and compatibility, with reduced boot time as a side effect. Although it can speed up startup in some cases, the main focus is on preventing problems and ensuring the system works well in the long term.

The idea behind the "sync probe" is to improve system stability and reliability. However, this forced synchronization can come at a cost in terms of performance.

Signed-off-by: TogoFire <togofire@mailfence.com>
Change-Id: I8fbdfacb518c71d661aeac9443caf2095b8dd604