Revert "printk: use buffer from the stack space"
This reverts commit 6d8902d1aff44c2a8d2e926e1dd66c969ba828ad.

Reason:
You may lose important information.
Risk of crashes.
Difficult to implement and debug.
The solution offers benefits, but there are risks.
The use of printk is not that frequent.
The proposal is an optimization, not a complete solution.
There are concerns about the amount of information in the logs.
It is necessary to find a balance between memory optimization and information preservation.

Signed-off-by: TogoFire <togofire@mailfence.com>
Change-Id: Ie42b1f7500ca3467a19c6e76250442c9bbabd4fd